fix: improve uint8 null handling and code robustness
Addresses several code quality and robustness issues in the skills system:

1. Added normalize_target_level helpers for consistent NULL_U8_SENTINEL handling
   - normalize_target_level_to_python: Converts Polars values (None/255) to Python (None/int)
   - normalize_target_level_to_polars: Converts Python values to Polars representation
   - Used in get_skill_training to prevent sentinel value bugs

2. Added defensive checks for Polars row iteration
   - get_skill_training now uses .get() with None checks for weight values
   - _sync_skills_to_legacy uses defensive casts with None checks
   - Added debug logging to _sync_skills_to_legacy for parity issue tracing

3. Documented Numba global array usage
   - Added comprehensive comment to APTITUDE_MULTIPLIERS explaining fragility
   - Updated numba_warmup() docstring to note it validates global array access
   - Ensures CI tests will catch compilation issues

4. Documented UsageWindow._decay float→uint32 behavior
   - Added detailed docstring explaining intentional floor operation
   - Clarifies that quick erosion of small counts is expected behavior
   - Documents alternatives if precise fractional tracking is needed

5. Added explicit dtypes to Polars Series creation
   - system.py now uses dtype=pl.UInt8 when creating final_levels Series
   - Improves clarity and prevents potential dtype coercion issues

All changes maintain backward compatibility and follow the project's
strict typing and documentation standards.

+++ skills/system.py
@@ -196,7 +196,9 @@ def _award_xp_impl(
         final_aptitudes = skills["aptitude"].to_numpy()
         final_levels = batch_calculate_levels(final_xp, final_aptitudes)
 
-        updates = updates.with_columns([pl.Series(final_levels).alias("level_new")])
+        updates = updates.with_columns(
+            [pl.Series(final_levels, dtype=pl.UInt8).alias("level_new")]
+        )
 
     # 9) Update skills_df atomically
     registry.skills_df = (
